fix: releaseVersion 64비트 전환
releaseVersion과 sync 포인터를 long/bigint 기준으로 맞추고, long 범위 초과 매니페스트·포인터 값의 처리되지 않은 오버플로 크래시 방지.

+++ src/GamePatchKit.Cli/PatchManifest.cs
@@ -13,7 +13,7 @@ internal sealed class PatchManifest
     public int SchemaVersion { get; init; } = CURRENT_SCHEMA_VERSION;
 
     [JsonProperty("releaseVersion", Required = Required.Always, Order = 1)]
-    public int ReleaseVersion { get; init; }
+    public long ReleaseVersion { get; init; }
 
     [JsonProperty("sourcePath", Required = Required.Always, Order = 2)]
     public string SourcePath { get; init; } = null!;
